+++ game_logic.h
@@ -32,6 +32,6 @@ bool is_legal_move(GameState& game_state, const uint32_t pawn_idx);
 
 bool any_pawn_left(GameState& game);
 
-int verify_game_state_after_move (GameState& game);
+uint8_t verify_game_state_after_move (GameState& game);
 
 #endif //KAYLES_GAME_LOGIC_H
